Switch to emotion/theme-ui, with a nasty patch to get no-js dark mode
Based on this example:
https://github.com/vercel/next.js/tree/ce5d9c858b70f647b532ec2f62ddde506af32b6a/examples/with-emotion-11

theme-ui contains a lot of stuff I was going to have to build on my own,
and emotion works almost identically to styled-components nowadays with a
slightly smaller package size.

Getting next-page-tester to work here was a nightmare - they only have a
styletron example and emotion/theme-ui are made up of like a dozen
packages, which would break if I isolated all of them.

I also hacked in a working dark mode that does not depend on JS